Deduplicate generateRunID into common.GenerateRunID
The PR's new-code duplicated-lines gate failed: migrate.go and
extract.go each carried their own byte-for-byte identical
generateRunID function (and wizard's had converged to the same
algorithm too, after a separate fix). These three hand-synced copies
predate this PR but only became "new" duplicated code once #542
touched all three bodies to add four-digit zero-padding.

Consolidate all three into common.GenerateRunID, and update every call
site (migrate.Run, migrate.RunReset, extract.prepareExtractDir,
wizard's extract/migrate retry loops) to use it. Also consolidates the
three near-identical test files into one authoritative
common/runid_test.go, alongside the existing RunIDAfter tests from
#542.
