+++ go/internal/common/runid.go
@@ -5,9 +5,45 @@
 package common
 
 import (
+	"fmt"
+	"os"
 	"strconv"
 	"strings"
+	"time"
 )
+
+// GenerateRunID returns an ISO-date-prefixed run ID (issue #108) for a
+// new run/extract under directory. Format: "YYYY-MM-DD-NNNN" where NNNN
+// is the next sequence number for the current day, zero-padded to four
+// digits so lexicographic and numeric ordering agree up to 9999 runs/day
+// (#542 — the previous two-digit padding let a 3-digit run number like
+// "-101" sort lexicographically *before* "-99").
+//
+// The next sequence number is the highest existing one for today, plus
+// one. An earlier (count-of-dirs + 1) approach broke once the numbering
+// had ANY gap — e.g. dirs -0010..-0019 with none below would yield
+// count=10, colliding with the existing -0011 and silently reusing its
+// task outputs. See the #359 follow-up regression report. This one
+// function backs migrate, extract, and wizard's run/extract ID
+// generation — they used to keep three hand-synced copies in sync,
+// which is exactly the kind of drift that let this wizard's copy fall
+// behind with the buggy count-based approach until #542.
+func GenerateRunID(directory string) string {
+	today := time.Now().UTC().Format("2006-01-02")
+	prefix := today + "-"
+	entries, _ := os.ReadDir(directory)
+	maxN := 0
+	for _, e := range entries {
+		if !e.IsDir() || !strings.HasPrefix(e.Name(), prefix) {
+			continue
+		}
+		n, err := strconv.Atoi(e.Name()[len(prefix):])
+		if err == nil && n > maxN {
+			maxN = n
+		}
+	}
+	return fmt.Sprintf("%s-%04d", today, maxN+1)
+}
 
 // RunIDAfter reports whether run/extract ID a is more recent than b.
 // Run IDs are "<date-prefix>-<NN>" where NN is zero-padded to a minimum
