+++ go/internal/extract/extract.go
@@ -12,8 +12,6 @@ import (
 	"os"
 	"path/filepath"
 	"sort"
-	"strconv"
-	"strings"
 	"sync"
 	"time"
 
@@ -215,7 +213,7 @@ func ListAllProjectKeys(ctx context.Context, cfg ExtractConfig) ([]string, error
 func prepareExtractDir(cfg ExtractConfig) (string, string, error) {
 	extractID := cfg.ExtractID
 	if extractID == "" {
-		extractID = generateRunID(cfg.ExportDirectory)
+		extractID = common.GenerateRunID(cfg.ExportDirectory)
 	}
 	extractDir := filepath.Join(cfg.ExportDirectory, extractID)
 	if err := os.MkdirAll(extractDir, 0o755); err != nil {
@@ -364,35 +362,6 @@ func detectEdition(ctx context.Context, raw *RawClient) (Edition, error) {
 		return EditionCommunity, err
 	}
 	return ParseEdition(body), nil
-}
-
-// generateRunID returns an ISO-date-prefixed extract ID (issue
-// #108). Format: "YYYY-MM-DD-NNNN", zero-padded to four digits so
-// lexicographic and numeric ordering agree up to 9999 runs/day
-// (#542). See migrate.generateRunID for the rationale — the two
-// helpers are deliberately kept in sync.
-func generateRunID(directory string) string {
-	today := time.Now().UTC().Format("2006-01-02")
-	prefix := today + "-"
-	entries, _ := os.ReadDir(directory)
-	maxN := 0
-	for _, e := range entries {
-		if !e.IsDir() {
-			continue
-		}
-		name := e.Name()
-		if !strings.HasPrefix(name, prefix) {
-			continue
-		}
-		n, err := strconv.Atoi(name[len(prefix):])
-		if err != nil {
-			continue
-		}
-		if n > maxN {
-			maxN = n
-		}
-	}
-	return fmt.Sprintf("%s-%04d", today, maxN+1)
 }
 
 // extractMeta groups the parameters for writeMetadata. Version stays as
