+++ go/internal/extract/extract_test.go
@@ -14,7 +14,6 @@ import (
 	"path/filepath"
 	"reflect"
 	"testing"
-	"time"
 )
 
 const (
@@ -870,68 +869,10 @@ func TestFetchAndWriteSingleWithResultKey(t *testing.T) {
 	}
 }
 
-// Regression for the runID-collision bug — twin of the migrate-side
-// test in migrate/migrate_test.go. Both helpers must use max+1 so
-// extract and migrate never alias onto an existing run dir when the
-// numbering has gaps.
-func TestGenerateRunID_HandlesNumberingGaps(t *testing.T) {
-	today := time.Now().UTC().Format("2006-01-02")
-
-	t.Run("empty directory yields -0001", func(t *testing.T) {
-		dir := t.TempDir()
-		got := generateRunID(dir)
-		want := today + "-0001"
-		if got != want {
-			t.Errorf("want %q, got %q", want, got)
-		}
-	})
-
-	t.Run("dirs -10..-19 yields -0020 (not -11 collision)", func(t *testing.T) {
-		dir := t.TempDir()
-		for i := 10; i <= 19; i++ {
-			if err := os.MkdirAll(filepath.Join(dir, fmt.Sprintf("%s-%02d", today, i)), 0o755); err != nil {
-				t.Fatalf("mkdir: %v", err)
-			}
-		}
-		got := generateRunID(dir)
-		want := today + "-0020"
-		if got != want {
-			t.Errorf("want %q (max+1), got %q", want, got)
-		}
-	})
-
-	t.Run("non-contiguous numbering still returns max+1", func(t *testing.T) {
-		dir := t.TempDir()
-		for _, n := range []int{1, 3, 7, 42} {
-			if err := os.MkdirAll(filepath.Join(dir, fmt.Sprintf("%s-%02d", today, n)), 0o755); err != nil {
-				t.Fatalf("mkdir: %v", err)
-			}
-		}
-		got := generateRunID(dir)
-		want := today + "-0043"
-		if got != want {
-			t.Errorf("want %q, got %q", want, got)
-		}
-	})
-
-	// #542 — the sequence number is zero-padded to four digits so
-	// that among IDs generated after this fix, lexicographic order
-	// keeps matching numeric order past the 99th run of a day (a
-	// pre-existing, non-padded "-99" dir from before this fix is a
-	// separate, unavoidable transition case handled by
-	// common.RunIDAfter instead, not by generateRunID's own output).
-	t.Run("crossing the 99->100 boundary still zero-pads to four digits", func(t *testing.T) {
-		dir := t.TempDir()
-		if err := os.MkdirAll(filepath.Join(dir, today+"-99"), 0o755); err != nil {
-			t.Fatalf("mkdir: %v", err)
-		}
-		got := generateRunID(dir)
-		want := today + "-0100"
-		if got != want {
-			t.Errorf("want %q, got %q", want, got)
-		}
-	})
-}
+// generateRunID moved to common.GenerateRunID (#542) — its own tests
+// now live in go/internal/common/runid_test.go, deduplicated across
+// the three packages (migrate, extract, wizard) that used to each
+// hand-maintain an identical copy of this function and its test.
 
 // #529: ListAllProjectKeys is the pre-flight, unfiltered project listing
 // transfer uses to resolve a --project_key regex pattern before the
