+++ go/internal/wizard/helpers_test.go
@@ -10,7 +10,6 @@ import (
 	"os"
 	"path/filepath"
 	"testing"
-	"time"
 
 	"github.com/sonar-solutions/sonar-migration-tool/internal/structure"
 )
@@ -326,34 +325,10 @@ func TestPtrStr(t *testing.T) {
 	}
 }
 
-func TestGenerateRunID(t *testing.T) {
-	dir := t.TempDir()
-	id := generateRunID(dir)
-	if id == "" {
-		t.Error("generateRunID should return non-empty string")
-	}
-	if id[len(id)-5:] != "-0001" {
-		t.Errorf("expected -0001 suffix, got %q", id)
-	}
-	// Issue #108 — the date prefix is ISO (YYYY-MM-DD), not the
-	// previous US format (MM-DD-YYYY). The ISO format starts with
-	// the four-digit year and sorts chronologically. We don't
-	// hard-code today's date (test would drift) but pin the shape:
-	// id must be at least YYYY-MM-DD-NNNN = 15 chars (#542 — the
-	// sequence number is zero-padded to four digits so lexicographic
-	// and numeric ordering agree up to 9999 runs/day), must start
-	// with today's UTC year, and the 5th character must be a hyphen.
-	want := time.Now().UTC().Format("2006")
-	if len(id) < 15 {
-		t.Fatalf("id too short for ISO format YYYY-MM-DD-NNNN: %q", id)
-	}
-	if id[:4] != want {
-		t.Errorf("id must start with current UTC year %q, got %q (full id %q)", want, id[:4], id)
-	}
-	if id[4] != '-' {
-		t.Errorf("id[4] must be '-' for ISO format, got %q (full id %q)", string(id[4]), id)
-	}
-}
+// generateRunID moved to common.GenerateRunID (#542) — its own tests
+// now live in go/internal/common/runid_test.go, deduplicated across
+// the three packages (migrate, extract, wizard) that used to each
+// hand-maintain an identical copy of this function and its test.
 
 func TestPhaseDisplayName(t *testing.T) {
 	tests := []struct {
